requirements.txt
altcha==0.2.0
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-allauth==65.10.0
idna==3.10
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
